IA-2876: move useRouter hook
